Removed Directory/Legacy mode
