+++ src/main/java/de/uniwue/controller/LineSegmentationController.java
@@ -15,9 +15,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.ModelAndView;
 
-import de.uniwue.helper.LineSegmentationDirectoryHelper;
 import de.uniwue.helper.LineSegmentationHelper;
-import de.uniwue.helper.LineSegmentationPageXMLHelper;
 
 /**
  * Controller class for line segmentation module
@@ -36,30 +34,13 @@ public class LineSegmentationController {
         if (GenericController.isSessionValid(session, response) == false)
             return null;
 
-        String processingMode = session.getAttribute("processingMode").toString();
-        
         // Keep a single helper object in session
         LineSegmentationHelper lineSegmentationHelper = (LineSegmentationHelper) session.getAttribute("lineSegmentationHelper");
-        if (lineSegmentationHelper == null || 
-        		(processingMode.equals("Directory") && lineSegmentationHelper instanceof LineSegmentationPageXMLHelper) ||
-        		(processingMode.equals("Pagexml") && lineSegmentationHelper instanceof LineSegmentationDirectoryHelper)) {
-
-        	// Select correct lineSegmentHelper for processingMode
-        	if(processingMode.equals("Directory")){
-				lineSegmentationHelper = new LineSegmentationDirectoryHelper(
-					session.getAttribute("projectDir").toString(),
-					session.getAttribute("imageType").toString(),
-					processingMode
-				);
-        	} else if(processingMode.equals("Pagexml")) {
-				lineSegmentationHelper = new LineSegmentationPageXMLHelper(
-					session.getAttribute("projectDir").toString(),
-					session.getAttribute("imageType").toString(),
-					processingMode
-				);
-        	} else {
-        		throw new IllegalArgumentException(String.format("Unknown processingMode %s", processingMode));
-        	}
+        if (lineSegmentationHelper == null) {
+            lineSegmentationHelper = new LineSegmentationHelper(
+                session.getAttribute("projectDir").toString(),
+                session.getAttribute("imageType").toString()
+            );
             session.setAttribute("lineSegmentationHelper", lineSegmentationHelper);
         }
         return lineSegmentationHelper;
