+++ src/main/java/de/uniwue/controller/OverviewController.java
@@ -42,8 +42,7 @@ public class OverviewController {
         if (overviewHelper == null) {
             overviewHelper = new OverviewHelper(
                 session.getAttribute("projectDir").toString(),
-                session.getAttribute("imageType").toString(),
-                session.getAttribute("processingMode").toString()
+                session.getAttribute("imageType").toString()
             );
             session.setAttribute("overviewHelper", overviewHelper);
         }
@@ -150,7 +149,6 @@ public class OverviewController {
     public @ResponseBody boolean checkDir(
                 @RequestParam("projectDir") String projectDir,
                 @RequestParam("imageType") String imageType,
-                @RequestParam("processingMode") String processingMode,
                 @RequestParam("resetSession") Boolean resetSession,
                 HttpSession session, HttpServletResponse response, HttpServletRequest request
             ) {
@@ -166,7 +164,6 @@ public class OverviewController {
         HttpSession newSession = request.getSession();
         newSession.setAttribute("projectDir", projectDir);
         newSession.setAttribute("imageType", imageType);
-        newSession.setAttribute("processingMode", processingMode);
         // Determine and add the name of the project to the session as well (to display on each page)
         String[] projectDirParts = projectDir.substring(0, projectDir.length() - 1).split(Pattern.quote(File.separator));
         newSession.setAttribute("projectName", projectDirParts[projectDirParts.length - 1]);
