+++ src/main/java/de/uniwue/controller/ResultGenerationController.java
@@ -38,8 +38,7 @@ public class ResultGenerationController {
         if (resultGenerationHelper == null) {
             resultGenerationHelper = new ResultGenerationHelper(
                 session.getAttribute("projectDir").toString(),
-                session.getAttribute("imageType").toString(),
-                session.getAttribute("processingMode").toString()
+                session.getAttribute("imageType").toString()
             );
             session.setAttribute("resultGenerationHelper", resultGenerationHelper);
         }
