+++ src/main/java/de/uniwue/controller/SegmentationController.java
@@ -33,8 +33,7 @@ public class SegmentationController {
         if (segmentationHelper == null) {
             segmentationHelper = new SegmentationHelper(
                 session.getAttribute("projectDir").toString(),
-                session.getAttribute("imageType").toString(),
-                session.getAttribute("processingMode").toString()
+                session.getAttribute("imageType").toString()
             );
             session.setAttribute("segmentationHelper", segmentationHelper);
         }
