+++ src/main/java/de/uniwue/controller/SegmentationPixelClassifierController.java
@@ -39,8 +39,7 @@ public class SegmentationPixelClassifierController {
         if (segmentationPixelClassifierHelper == null) {
             segmentationPixelClassifierHelper = new SegmentationPixelClassifierHelper(
                 session.getAttribute("projectDir").toString(),
-                session.getAttribute("imageType").toString(),
-                session.getAttribute("processingMode").toString()
+                session.getAttribute("imageType").toString()
             );
             session.setAttribute("segmentationPixelClassifier", segmentationPixelClassifierHelper);
         }
