+++ src/main/java/de/uniwue/feature/ProcessStateCollector.java
@@ -28,22 +28,15 @@ public class ProcessStateCollector {
     private String imageType;
 
     /**
-     * Processing structure of the project
-     * Possible values: { Directory, Pagexml }
-     */
-    private String processingMode;
-
-    /**
      * Constructor
      *
      * @param projConf Project configuration object
      * @param projectImageType Type of the project (gray, binary)
      * @param imageType Image type of the project
      */
-    public ProcessStateCollector(ProjectConfiguration projConf, String imageType, String processingMode) {
+    public ProcessStateCollector(ProjectConfiguration projConf, String imageType) {
         this.projConf = projConf;
         this.imageType = imageType;
-        this.processingMode = processingMode;
     }
 
     /**
@@ -84,46 +77,6 @@ public class ProcessStateCollector {
         return true;
     }
 
-    /**
-     * Checks whether the required image files in "OCR/Pages/{pageId}" directory exist or not
-     *
-     * @param pageID Identifier of the page (e.g 0002,0003)
-     * @param includePseg Determines if the ".pseg.png" file should be checked as well (for line segmentation only)
-     * @return Information if the required image files exist
-     */
-    public boolean existPageDirImageFiles(String pageId, boolean includePseg) {
-        File pageDir = new File(projConf.PAGE_DIR + pageId);
-        if (!pageDir.exists())
-            return false;
-
-        // Identify plain region images (REGION_ID.png)
-        File[] plainRegionImgFiles = pageDir.listFiles((d, name) ->
-            name.endsWith(projConf.IMG_EXT) &&
-            !name.endsWith(projConf.BINR_IMG_EXT) &&
-            !name.endsWith(projConf.GRAY_IMG_EXT) &&
-            !name.endsWith(projConf.PSEG_EXT)
-        );
-
-        if (plainRegionImgFiles.length == 0)
-            return false;
-
-        // Check that for every region file a ".bin.png" and ".nrm.png exists (successful region extraction)
-        // If required, check that for every region file a ".pseg.png" exists (successful line segmentation)
-        for (File imgFile : plainRegionImgFiles) {
-            if (!new File(FilenameUtils.removeExtension(imgFile.getAbsolutePath()) + projConf.BINR_IMG_EXT).exists())
-                return false;
-
-            if (!new File(FilenameUtils.removeExtension(imgFile.getAbsolutePath()) + projConf.GRAY_IMG_EXT).exists())
-                return false;
-
-            if (includePseg == true) {
-                if (!new File(FilenameUtils.removeExtension(imgFile.getAbsolutePath()) + projConf.PSEG_EXT).exists())
-                    return false;
-            }
-        }
-
-        return true;
-    }
 
     /**
      * Checks whether the required textlines for the image files exist or not.
@@ -165,31 +118,14 @@ public class ProcessStateCollector {
     }
 
     /**
-     * Determines the "RegionExtraction" process state of a given page
-     *
-     * @param pageID Identifier of the page (e.g 0002,0003)
-     * @return "RegionExtraction" state of the page
-     */
-    public boolean regionExtractionState(String pageId) {
-        // Check that for every region file a ".bin.png" and ".nrm.png exists (successful region extraction)
-        return existPageDirImageFiles(pageId, false);
-    }
-
-    /**
      * Determines the "LineSegmentation" process state of a given page
      *
      * @param pageID Identifier of the page (e.g 0002,0003)
      * @return "LineSegmentation" state of the page
      */
     public boolean lineSegmentationState(String pageId) {
-    	if(processingMode.equals("Directory")) {
-			// Check that for every region file a ".bin.png" and ".nrm.png exists (successful region extraction)
-			// Check that for every region file a ".pseg.png" exists (successful line segmentation)
-			return existPageDirImageFiles(pageId, true);
-    	} else {
-    		// Check whether lines in the PageXML files exist
-    		return existLines(pageId);
-    	}
+        // Check whether lines in the PageXML files exist
+        return existLines(pageId);
     }
 
     /**
@@ -199,56 +135,30 @@ public class ProcessStateCollector {
      * @return "Recognition" state of the page
      */
     public boolean recognitionState(String pageId) {
-    	if(processingMode.equals("Directory")){
-    		// Check directory for recognition txts
-			File pageDir = new File(projConf.PAGE_DIR + pageId);
-			if (!pageDir.exists())
-				return false;
+        // Check pagexml for TextEquivs
+        File pageXML = new File(projConf.OCR_DIR + pageId + projConf.CONF_EXT);
+        if (!pageXML.exists())
+            return false;
 
-			File[] lineSegmentDirectories = pageDir.listFiles(File::isDirectory);
-			if (lineSegmentDirectories.length == 0) {
-				return false;
-			}
+        // Easy and fast check if an end tag of a TextLine exists 
+        // Does not check if the xml is valid
+        try (FileInputStream fis = new FileInputStream(pageXML)){
+            byte[] data = new byte[(int) pageXML.length()];
+            fis.read(data);
+            fis.close();
+            String pageXMLContent = new String(data, "UTF-8");
+            // Test for TextEquiv with index higher 0 (GT)
+            Pattern p = Pattern.compile("\\<TextEquiv[^>]+?index=\"[^0]\"[^>]*?\\>");
+            Matcher matcher = p.matcher(pageXMLContent);
 
-			for (File dir : lineSegmentDirectories) {
-				File[] lineSegmentImgFiles = dir.listFiles((d, name) -> name.endsWith(projConf.getImageExtensionByType(imageType)));
-				// Check that for every line segmenation file a ".txt" exists (successful recognition)
-				for (File imgFile : lineSegmentImgFiles) {
-					// Call removeExtension twice, due to "Binary"|"Gray" image extensions (".bin.png"|".nrm.png")
-					if (!new File(FilenameUtils.removeExtension(FilenameUtils.removeExtension(imgFile.getAbsolutePath()))
-							 + projConf.REC_EXT).exists()) {
-						return false;
-					}
-				}
-			}
-
-			return true;
-    	} else {
-    		// Check pagexml for TextEquivs
-			File pageXML = new File(projConf.OCR_DIR + pageId + projConf.CONF_EXT);
-			if (!pageXML.exists())
-				return false;
-
-			// Easy and fast check if an end tag of a TextLine exists 
-			// Does not check if the xml is valid
-			try (FileInputStream fis = new FileInputStream(pageXML)){
-				byte[] data = new byte[(int) pageXML.length()];
-				fis.read(data);
-				fis.close();
-				String pageXMLContent = new String(data, "UTF-8");
-				// Test for TextEquiv with index higher 0 (GT)
-				Pattern p = Pattern.compile("\\<TextEquiv[^>]+?index=\"[^0]\"[^>]*?\\>");
-				Matcher matcher = p.matcher(pageXMLContent);
-
-				return matcher.find();
-			} catch (UnsupportedEncodingException e) {
-				return false;
-			} catch (FileNotFoundException e) {
-				return false;
-			} catch (IOException e) {
-				return false;
-			}
-    	}
+            return matcher.find();
+        } catch (UnsupportedEncodingException e) {
+            return false;
+        } catch (FileNotFoundException e) {
+            return false;
+        } catch (IOException e) {
+            return false;
+        }
     }
 
     /**
@@ -258,47 +168,33 @@ public class ProcessStateCollector {
      * @return "result" state of the page
      */
     public boolean groundTruthState(String pageId) {
-        if(processingMode.equals("Directory")) {
-        	// Check if any file has Ground Truth
-            File[] directories = new File(projConf.PAGE_DIR + pageId).listFiles(File::isDirectory);
-            if (directories != null && directories.length != 0) {
-                for(File dir : directories) {
-                    File[] GtFiles = dir.listFiles((d, name) -> name.endsWith(projConf.GT_EXT));
-                    if(GtFiles.length > 0) {
-                    	return true;
-                    }
-                }
+        // Check pagexml for TextEquivs
+        File pageXML = new File(projConf.OCR_DIR + pageId + projConf.CONF_EXT);
+        if (!pageXML.exists())
+            return false;
+
+        // Easy and fast check if an end tag of a TextLine exists 
+        // Does not check if the xml is valid
+        try (FileInputStream fis = new FileInputStream(pageXML)){
+            byte[] data = new byte[(int) pageXML.length()];
+            fis.read(data);
+            fis.close();
+            String pageXMLContent = new String(data, "UTF-8");
+            
+            Pattern p = Pattern.compile("\\<TextEquiv[^>]+?index=\"0\"");
+            Matcher matcher = p.matcher(pageXMLContent);
+
+            if(matcher.find()) {
+                return true;
             }
-			return false;
-        } else {
-    		// Check pagexml for TextEquivs
-			File pageXML = new File(projConf.OCR_DIR + pageId + projConf.CONF_EXT);
-			if (!pageXML.exists())
-				return false;
-
-			// Easy and fast check if an end tag of a TextLine exists 
-			// Does not check if the xml is valid
-			try (FileInputStream fis = new FileInputStream(pageXML)){
-				byte[] data = new byte[(int) pageXML.length()];
-				fis.read(data);
-				fis.close();
-				String pageXMLContent = new String(data, "UTF-8");
-				
-				Pattern p = Pattern.compile("\\<TextEquiv[^>]+?index=\"0\"");
-				Matcher matcher = p.matcher(pageXMLContent);
-
-				if(matcher.find()) {
-					return true;
-				}
-			} catch (UnsupportedEncodingException e) {
-				return false;
-			} catch (FileNotFoundException e) {
-				return false;
-			} catch (IOException e) {
-				return false;
-			}
-			return false;
+        } catch (UnsupportedEncodingException e) {
+            return false;
+        } catch (FileNotFoundException e) {
+            return false;
+        } catch (IOException e) {
+            return false;
         }
+        return false;
     }
 
     /**
