+++ src/main/java/de/uniwue/helper/DespecklingHelper.java
@@ -41,11 +41,10 @@ public class DespecklingHelper {
      *
      * @param projectDir Path to the project directory
      * @param projectImageType Type of the project (binary, gray)
-     * @param processingMode Processing structure of the project (Directory, Pagexml)
      */
-    public DespecklingHelper(String projectDir, String projectImageType, String processingMode) {
+    public DespecklingHelper(String projectDir, String projectImageType) {
         projConf = new ProjectConfiguration(projectDir);
-        procStateCol = new ProcessStateCollector(projConf, projectImageType, processingMode);
+        procStateCol = new ProcessStateCollector(projConf, projectImageType);
     }
 
     /**
