+++ src/main/java/de/uniwue/helper/PreprocessingHelper.java
@@ -24,12 +24,6 @@ public class PreprocessingHelper {
      * Image Type of the Project
      */
     private String projectImageType;
-
-    /**
-     * Processing structure of the project
-     * Possible values: { Directory, Pagexml }
-     */
-    private String processingMode;
     
     /**
      * Object to determine process states
@@ -69,14 +63,12 @@ public class PreprocessingHelper {
      *
      * @param projectDir Path to the project directory
      * @param projectImageType Type of the project (binary, gray)
-     * @param processingMode Processing structure of the project (Directory, Pagexml)
      */
-    public PreprocessingHelper(String projectDir, String projectImageType, String processingMode) {
+    public PreprocessingHelper(String projectDir, String projectImageType) {
         projConf = new ProjectConfiguration(projectDir);
-        procStateCol = new ProcessStateCollector(projConf, projectImageType, processingMode);
+        procStateCol = new ProcessStateCollector(projConf, projectImageType);
         processHandler = new ProcessHandler();
         this.projectImageType = projectImageType;
-        this.processingMode = processingMode;
     }
 
     /**
@@ -221,7 +213,7 @@ public class PreprocessingHelper {
         }
 
         // Delete Despeckled images as well (they are generated from Binary images)
-        DespecklingHelper despecklingHelper = new DespecklingHelper(projConf.PROJECT_DIR, projectImageType, processingMode);
+        DespecklingHelper despecklingHelper = new DespecklingHelper(projConf.PROJECT_DIR, projectImageType);
         despecklingHelper.deleteOldFiles(pageIds);
     }
