+++ src/main/java/de/uniwue/helper/SegmentationHelper.java
@@ -19,11 +19,6 @@ public class SegmentationHelper {
      */
     private String projectImageType;
 
-    /**
-     * Processing structure of the project
-     * Possible values: { Directory, Pagexml }
-     */
-    private String processingMode;
     
     /**
      * Object to determine process states
@@ -35,13 +30,11 @@ public class SegmentationHelper {
      *
      * @param projectDir Path to the project directory
      * @param projectImageType Type of the project (binary,gray)
-     * @param processingMode Processing structure of the project (Directory, Pagexml)
      */
-    public SegmentationHelper(String projDir, String projectImageType, String processingMode) {
-    	this.processingMode = processingMode;
+    public SegmentationHelper(String projDir, String projectImageType) {
         this.projectImageType = projectImageType;
         projConf = new ProjectConfiguration(projDir);
-        procStateCol = new ProcessStateCollector(projConf, projectImageType, processingMode);
+        procStateCol = new ProcessStateCollector(projConf, projectImageType);
     }
 
     /**
@@ -67,10 +60,6 @@ public class SegmentationHelper {
     public void deleteOldFiles(List<String> pageIds) throws IOException {
         if (!new File(projConf.OCR_DIR).exists())
             return;
-
-        // Delete all files created by subsequent processes to preserve data integrity
-        RegionExtractionHelper regionExtracorHelper = new RegionExtractionHelper(projConf.PROJECT_DIR, this.projectImageType, processingMode);
-        regionExtracorHelper.deleteOldFiles(pageIds);
 
         // Delete image and PageXML files
         for (String pageId : pageIds) {
