+++ src/main/java/de/uniwue/helper/TrainingHelper.java
@@ -30,11 +30,6 @@ public class TrainingHelper {
      */
     private String projectImageType;
 
-    /**
-     * Processing structure of the project
-     * Possible values: { Directory, Pagexml }
-     */
-    private String processingMode;
     
     /**
      * Object to access project configuration
@@ -61,13 +56,11 @@ public class TrainingHelper {
      *
      * @param projectDir Path to the project directory
      * @param projectImageType Type of the project (binary, gray)
-     * @param processingMode Processing structure of the project (Directory, Pagexml)
      */
-    public TrainingHelper(String projectDir, String projectImageType, String processingMode) {
+    public TrainingHelper(String projectDir, String projectImageType) {
         projConf = new ProjectConfiguration(projectDir);
         processHandler = new ProcessHandler();
         this.projectImageType = projectImageType;
-        this.processingMode = processingMode;
     }
 
     /**
@@ -99,7 +92,7 @@ public class TrainingHelper {
      */
     public List<String> getImagesWithGt(String projectImageType) throws IOException {
         ArrayList<String> imagesWithGt = new ArrayList<String>();
-        final String gtExt = processingMode.equals("Pagexml") ? projConf.CONF_EXT : projConf.GT_EXT;
+        final String gtExt = projConf.CONF_EXT;
 
         // Add custom models to map
 		Files.walk(Paths.get(projConf.PAGE_DIR))
@@ -206,10 +199,8 @@ public class TrainingHelper {
 
         command.add("--no_progress_bars");
 
-        if(processingMode.equals("Pagexml")) {
-        	command.add("--dataset");
-        	command.add("PAGEXML");
-        }
+        command.add("--dataset");
+        command.add("PAGEXML");
 
         processHandler = new ProcessHandler();
         processHandler.setFetchProcessConsole(true);
