+++ src/main/webapp/WEB-INF/tags/body.tag
@@ -45,32 +45,12 @@
                 </ul>
             </li>
 
-			<c:choose>
-				<c:when test='${(not empty processingMode) && (processingMode == "Pagexml")}'>
-				</c:when>
-				<c:otherwise>
-            <li><a href="RegionExtraction">Region Extraction</a></li>
-				</c:otherwise>
-			</c:choose>
             <li><a href="LineSegmentation">Line Segmentation</a></li>
             <li><a href="Recognition">Recognition</a></li>
-			<c:choose>
-				<c:when test='${(not empty processingMode) && (processingMode == "Pagexml")}'>
 			<li><a href="GroundTruthProductionLarex">Ground Truth Production</a></li>
-				</c:when>
-				<c:otherwise>
-            <li><a href="/GTC_Web?gtcDir=${fn:replace(projectDir, '\\', '/')}processing&dirType=pages" target="_blank">Ground Truth Production</a></li>
-				</c:otherwise>
-			</c:choose>
             <li><a href="Training">Training</a></li>
             <li><a href="Evaluation">Evaluation</a></li>
-			<c:choose>
-				<c:when test='${(not empty processingMode) && (processingMode == "Pagexml")}'>
             <li><a href="PostCorrectionLarex">Post Correction</a></li>
-				</c:when>
-				<c:otherwise>
-				</c:otherwise>
-			</c:choose>
             <li><a href="ResultGeneration">Result Generation</a></li>
             <li class="nav-separator"></li>
         </ul>
